Added GPL license
You cannot have closed source copies of this repository. All further work on this repo has to be open source.
Please, contact me if you want to work on this so I can add you as maintainer.
Since this repo never had a license, you ARE NOT FREE to have previous versions of this code as private. lootbot is now licensed under GPLv3 and all previous releases too.